fastapi_backend/requirements.txt
# FastAPI Backend Requirements

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# MathPix SDK
python-mathpix==0.4

# Async support
asyncio==3.4.3

# File handling
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Existing project dependencies
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0
TexSoup>=0.3.1
pylatexenc>=2.10

# Document processing (from doc_edit)
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Utilities
rich>=13.0.0
loguru>=0.7.0
click>=8.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI

# CORS
python-cors==1.0.0
